// frontend/src/types/generated/BindingRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EventVariable } from "./EventVariable";
import type { LabelValue } from "./LabelValue";
import type { ObjectVariable } from "./ObjectVariable";
import type { ViolationReason } from "./ViolationReason";

export type BindingRow = { objects: Array<[ObjectVariable, string]>, events: Array<[EventVariable, string]>, labels: Array<[string, LabelValue]>, violation: ViolationReason | null, };