// frontend/src/types/generated/LabelValue.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type LabelValue = { "type": "string", "value": string } | { "type": "int", "value": bigint } | { "type": "float", "value": number } | { "type": "bool", "value": boolean } | { "type": "null" };